luci-app-nlbwmon: Add translation context for Download/Upload terms
Added a translation context for the terms 'Download' and 'Upload'
used in this application. The reason is that these terms are
translated differently in Russian depending on the context.

If the terms are used in the sense of an action (download file
or upload file) then the translations should be as follows:
- Download = скачать
- Upload   = загрузить

But if they are used in the context of a traffic counter (like
in this application) the correct translations are:
- Download = скачано
- Upload   = загружено

A translation context has been added to all affected PO files
to avoid breaking existing translations. The Russian translation
in PO file is fixed to comply with the used context.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>
(cherry picked from commit a06a1bc7d4a8bc5df513dcfadc1069595ab48d87)