luci-app-softethervpn: update provided links for SoftEtherVPN clients
I updated the link for downloading Windows client to download the latest
currently available version instead of using beta version, which is not
good to use in production.

Also, the macOS link led to the OpenVPN server, which was wrong. It should
be a client. It was updated to the latest version as well.

[rebased to latest Lean's source]
Signed-off-by: Tianling Shen <cnsztl@immortalwrt.org>
(cherry picked from commit a93d4534eaf35e6b0d7fa0e088233e1d746e2c23)