luci-mod-status: display full HE rate information
Associated Stations on the status page should also show full HE rate
information which was added by
commit 49ae93eb9a ("luci-mod-network: add 802.11ax HE support")

However, it was only added to the Wireless page and not to the summary
on the Status page. Add it there as well.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>
(cherry picked from commit 14fbaa5266287a624d5203704596dda652d0bb68)