luci-app-attendedsysupgrade: fix missing efi path
If running on a non EFI system, the file `/sys/firmware/efi` is not
available and therefore results in an error of `fs.stat`. Wrap it with
`L.resolveDefault` to avoid the error message and make it work on
non-EFI devices again.

Signed-off-by: Paul Spooren <mail@aparcar.org>
(cherry picked from commit 03aceaafdc9551134d91955c8b6311d1838e71d3)